fix(rate-of-closure): type capability attempt count

+++ src/shared/python/swing_sim/flight/_capability_observation_runtime.py
@@ -41,7 +41,7 @@ class _OptimizationContext:
 
     @property
     def total_count(self) -> int:
-        return self.request.candidate_budget * self.request.ensemble_size
+        return int(self.request.candidate_budget * self.request.ensemble_size)
 
 
 @dataclass(frozen=True)
